test(se050): refactor and add tests for ExtraData parsing
