+++ se050/Cargo.toml
@@ -14,6 +14,7 @@ rust-version.workspace = true
 clap.workspace = true
 color-eyre.workspace = true
 derive_more = { workspace = true, features = ["from", "into"] }
+hex-literal.workspace = true
 orb-telemetry.workspace = true
 owo-colors.workspace = true
 thiserror.workspace = true
